pyproject.toml
[project]
name = "ec-cot-experiment"
version = "0.1.0"
description = "Evidence-Coverage Chain-of-Thought prompt tuning experiment"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.0",
    "wandb>=0.16.0",
    "google-generativeai>=0.3.0",
    "datasets>=2.14.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "tqdm>=4.66.0",
    "matplotlib>=3.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"